requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
requests>=2.31.0
PyYAML>=6.0.1
feedparser>=6.0.11
praw>=7.7.1
markdown>=3.6
